add Dockerfile to customize Jetify base image
Add a Dockerfile that builds upon the `jetify-sandbox` base image by
replacing the default editor with github.com/stencila/writer. This
serves as an example of how to customize the image that runs in Jetify
sandboxes.

The image should be built with `--platform linux/amd64` to ensure that
it is compatible with the VMs, which are x86 machines.

To build the image (this may take over an hour on a slower machine):

	docker build --platform linux/amd64 --tag <image-name> .

To run the Code server locally on port 8080:

	docker run --rm --privileged --platform linux/amd64 --publish 8080:80 <image-name>

To start a shell (and skip starting the server) for debugging:

	docker run --rm --privileged --platform linux/amd64 --publish 8080:80 --entrypoint /bin/bash --interactive --tty <image-name>

A few tips on modifying the image:

- The container entrypoint is `/google/scripts/entrypoint.sh`.
- Startup scripts are in `/etc/workstation-startup.d/*`. They run in
  alphabetical order and the last script starts the editor server.
- The editor lives in /opt/code-oss. Some other startup scripts assume
  this path, so be careful if renaming or moving it.
- The user's home directory (by default `/home/user`) does not exist
  until the first login. Add files to `/etc/skel` (such as .bashrc or VS
  Code settings) and they will be automatically copied into the user's
  home directory when it's created.